# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Drogaria",
        "description": "Esta documentação descreve os endpoints disponíveis para integração com o sistema web da Drogaria, desenvolvido em Laravel. A API pode ser utilizada por qualquer linguagem de programação capaz de realizar requisições HTTP.",
        "contact": {
            "name": "Bruno",
            "email": "bruno.facomp2016@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar todos os produtos",
                "operationId": "68de3067634ab4598e5d2fe1adc9b3da",
                "responses": {
                    "200": {
                        "description": "Lista de produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "codigo": {
                                                "type": "string",
                                                "example": "ABC123"
                                            },
                                            "descricao": {
                                                "type": "string",
                                                "example": "Produto Exemplo"
                                            },
                                            "preco": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 99.99
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-05-20T14:30:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-05-20T14:30:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Criar um novo produto",
                "operationId": "7c186445d191eabc09f48da5cb16782f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codigo",
                                    "descricao",
                                    "preco"
                                ],
                                "properties": {
                                    "codigo": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Produto Exemplo"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "codigo": {
                                            "type": "string",
                                            "example": "ABC123"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Produto Exemplo"
                                        },
                                        "preco": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 99.99
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-20T14:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-20T14:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/produtos/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibir um produto específico",
                "operationId": "358d7f18152ccef1268ce3edd2bb7f6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "codigo": {
                                            "type": "string",
                                            "example": "ABC123"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Produto Exemplo"
                                        },
                                        "preco": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 99.99
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-20T14:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-20T14:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar um produto",
                "operationId": "7b8dd7116a9918178331971ec5cca4b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo": {
                                        "type": "string",
                                        "example": "XYZ999"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Produto Atualizado"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 149.9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "codigo": {
                                            "type": "string",
                                            "example": "XYZ999"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Produto Atualizado"
                                        },
                                        "preco": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 149.9
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-20T14:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-20T14:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Deletar um produto",
                "operationId": "3d951d710d523eca6d888cbf7f3c698c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto removido com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        },
...
    },
...
}